refactor: apply notice modal
